requirement.txt
bokeh
websocket-client